init
------------------------------------------------------------------------------
- feat : 새로운 Feature가 추가
- fix : 버그 수정
- refactor : 기존 코드를 refactoring
- docs : document 수정 사항
- style : 비즈니스 로직에는 수정이 없으나 코드 스타일이 변경
- test : test code를 추가하거나 수정
- build : 빌드 스크립트 수정
- ci : ci 구성 파일 및 스크립트 변경
- chore : 기타 수정 사항
------------------------------------------------------------------------------

More detailed explanatory text, if necessary. Wrap it to about 72
characters or so. In some contexts, the first line is treated as the
subject of the commit and the rest of the text as the body. The
blank line separating the summary from the body is critical (unless
you omit the body entirely); various tools like `log`, `shortlog`
and `rebase` can get confused if you run the two together.

Explain the problem that this commit is solving. Focus on why you
are making this change as opposed to how (the code explains that).
Are there side effects or other unintuitive consequenses of this
change? Here's the place to explain them.

Further paragraphs come after blank lines.

 - Bullet points are okay, too

 - Typically a hyphen or asterisk is used for the bullet, preceded
   by a single space, with blank lines in between, but conventions
   vary here

If you use an issue tracker, put references to them at the bottom,
like this:

Resolves: #123
See also: #456, #789